requirements.txt
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-nose==1.4.6
djangorestframework==3.9.0
docutils==0.14
drftest==1.0.1
flake8==3.6.0
idna==2.8
mccabe==0.6.1
nose==1.3.7
pkginfo==1.4.2
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.0
pytz==2018.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.8.0
six==1.12.0
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
